+++ packages/integration/src/processVanillaFile.ts
@@ -1,5 +1,4 @@
 import { FileScope, Adapter } from '@vanilla-extract/css';
-import { setAdapter } from '@vanilla-extract/css/adapter';
 import { transformCss } from '@vanilla-extract/css/transformCss';
 // @ts-expect-error
 import evalCode from 'eval';
@@ -75,21 +74,22 @@ export function processVanillaFile({
     getIdentOption: () => identOption,
   };
 
-  setAdapter(cssAdapter);
-
   const currentNodeEnv = process.env.NODE_ENV;
-
-  const sourceWithBoundLoaderInstance = `require('@vanilla-extract/css/adapter').setAdapter(__adapter__);${source};`;
 
   // Vite sometimes modifies NODE_ENV which causes different versions (e.g. dev/prod) of vanilla packages to be loaded
   // This can cause CSS to be bound to the wrong instance, resulting in no CSS output
   // To get around this we set the NODE_ENV back to the original value ONLY during eval
   process.env.NODE_ENV = originalNodeEnv;
 
+  const adapterBoundSource = `
+    require('@vanilla-extract/css/adapter').setAdapter(__adapter__);
+    ${source}
+  `;
+
   const evalResult = evalCode(
-    sourceWithBoundLoaderInstance,
+    adapterBoundSource,
     filePath,
-    { console, __adapter__: cssAdapter, process },
+    { console, process, __adapter__: cssAdapter },
     true,
   );
 
@@ -118,6 +118,20 @@ export function processVanillaFile({
 
     cssImports.push(virtualCssFilePath);
   }
+
+  // We run this code inside eval as jest seems to create a difrerent instance of the adapter file
+  // for requires executed within the eval and all CSS can be lost.
+  evalCode(
+    `const { removeAdapter } = require('@vanilla-extract/css/adapter');
+    // Backwards compat with older versions of @vanilla-extract/css
+    if (removeAdapter) {
+      removeAdapter();
+    }
+  `,
+    filePath,
+    { console, process },
+    true,
+  );
 
   const unusedCompositions = composedClassLists
     .filter(({ identifier }) => !usedCompositions.has(identifier))