+++ test-helpers/src/index.ts
@@ -1,5 +1,4 @@
 export * from './startFixture';
-export * from './getNodeStyles';
 export * from './getStylesheet';
 
 export const getTestNodes = (fixture: string) =>